Add figaro gem ie 12-factor-app methodology
We want to move towards using ENV variables for tokens.

This commit is a partial step towards that goal. Tokens will be set in
application.yml and through Figaro.gem, that will be loaded into ENV
constant.

The longer and much better solution is to no longer keep tokens in the
checked in & public git repository. Hopefully we can shift towards those
ENV variables being deployed via Chef or Puppet or manually on server.